Refactor download CLI command
 - "datum describe-downloads" -> "datum download describe"
 - "datum download" -> "datum download get"
 - Refactor datum download cli test to use pytest framework
