+++ datumaro/cli/util/__init__.py
@@ -40,6 +40,27 @@ class MultilineFormatter(argparse.HelpFormatter):
             multiline_text += formatted_paragraph
         return multiline_text
 
+    def _format_action(self, action):
+        """
+        This requires to remove an ugly curly braced list generated by add_subparsers().
+        Please see the following example.
+
+        <w/o this patch>
+            Commands:
+            {get,describe}
+                get           Download a publicly available dataset
+                describe      Print information about downloadable datasets
+
+        <w/ this patch>
+            Commands:
+                get           Download a publicly available dataset
+                describe      Print information about downloadable datasets
+        """
+        parts = super()._format_action(action)
+        if action.nargs == argparse.PARSER:
+            parts = "\n".join(parts.split("\n")[1:])
+        return parts
+
 
 def required_count(nmin=0, nmax=0):
     assert 0 <= nmin and 0 <= nmax and nmin or nmax
